# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Algorithmic Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get trade log
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: strategy
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [open, closed, cancelled, null]
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeListResponse"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get single trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
  /trades/{id}/notes:
    patch:
      operationId: updateTradeNotes
      tags: [trades]
      summary: Update trade notes and tags
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTradeNotesBody"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
  /trades/calendar:
    get:
      operationId: getTradeCalendar
      tags: [trades]
      summary: Get daily trade summary for calendar view
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"
...
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get strategy details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "404":
          description: Not found

    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update strategy parameters
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStrategyBody"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}/toggle:
    post:
      operationId: toggleStrategy
      tags: [strategies]
      summary: Toggle strategy active state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        status:
          type: string
          enum: [open, closed, cancelled]
        strategy:
          type: string
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        quantity:
          type: number
        pnl:
          type: ["number", "null"]
        pnlPct:
          type: ["number", "null"]
        fees:
          type: number
        slippage:
          type: number
        mae:
          type: ["number", "null"]
        mfe:
          type: ["number", "null"]
        openedAt:
          type: string
        closedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        orderType:
          type: string
          enum: [market, limit]
        zScore:
          type: ["number", "null"]
        sentimentMultiplier:
          type: ["number", "null"]
      required:
        - id
        - symbol
        - side
        - status
        - strategy
        - entryPrice
        - exitPrice
        - quantity
        - pnl
        - pnlPct
        - fees
        - slippage
        - mae
        - mfe
        - openedAt
        - closedAt
        - notes
        - tags
        - stopLoss
        - takeProfit
        - orderType
        - zScore
        - sentimentMultiplier
    TradeListResponse:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
      required: [trades, total]
    UpdateTradeNotesBody:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    CalendarDay:
      type: object
      properties:
        date:
          type: string
        pnl:
          type: number
        trades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
      required: [date, pnl, trades, wins, losses, winRate]
...
    Strategy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [mean_reversion, momentum, statistical]
        active:
          type: boolean
        symbols:
          type: array
          items:
            type: string
        parameters:
          type: object
          additionalProperties: true
        riskPct:
          type: number
        description:
          type: string
        winRate:
          type: number
        totalTrades:
          type: integer
        pnl:
          type: number
      required:
        - id
        - name
        - type
        - active
        - symbols
        - parameters
        - riskPct
        - description
        - winRate
        - totalTrades
        - pnl
    UpdateStrategyBody:
      type: object
      properties:
        active:
          type: boolean
        riskPct:
          type: number
        parameters:
          type: object
          additionalProperties: true
        symbols:
          type: array
          items:
            type: string
...